feat: moves parseDatabaseUrl function to utility module
